+++ Sources/PowerSync/Protocol/db/CrudBatch.swift
@@ -13,3 +13,12 @@ public protocol CrudBatch {
     /// `writeCheckpoint` is optional.
     func complete(writeCheckpoint: String?) async throws
 }
+
+public extension CrudBatch {
+    /// Call to remove the changes from the local queue, once successfully uploaded.
+    func complete() async throws {
+        try await self.complete(
+            writeCheckpoint: nil
+        )
+    }
+}