+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.1.1 (unreleased)
+
+* Improved `CrudBatch` and `CrudTransaction` `complete` function extensions. Developers no longer need to specify `nil` as an argument for `writeCheckpoint` when calling `CrudBatch.complete`. The base `complete` functions still accept an optional `writeCheckpoint` argument if developers use custom write checkpoints. 
+``` diff
+guard let finalBatch = try await powersync.getCrudBatch(limit: 100) else {
+  return nil
+}
+- try await batch.complete(writeCheckpoint: nil)
++ try await batch.complete()
+```
+
 ## 1.1.0
 
 * Add sync progress information through `SyncStatusData.downloadProgress`.